requirements.txt
asgiref==3.4.1
attrs==21.2.0
black==21.9b0
boto3==1.18.42
botocore==1.21.42
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-storages==1.11.1
djangorestframework==3.12.4
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==8.13.2
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
mock==4.0.3
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
